+++ src/cli/base64.rs
@@ -1,0 +1,66 @@
+use clap::Parser;
+use std::fmt;
+use std::str::FromStr;
+
+use super::validate_path;
+
+#[derive(Debug, Parser)]
+pub enum Base64SubCommand {
+    #[command(name = "encode", about = "将文件或字符串编码为base64")]
+    Encode(Base64EncodeOpts),
+    #[command(name = "decode", about = "将base64编码的字符串或文件解码")]
+    Decode(Base64DecodeOpts),
+}
+
+#[derive(Debug, Parser)]
+pub struct Base64EncodeOpts {
+    #[arg(short, long, value_parser = validate_path, default_value = "-")]
+    pub input: String,
+    #[arg(long, value_parser = parse_base64_format, default_value = "standard")]
+    pub format: Base64Format,
+}
+
+#[derive(Debug, Parser)]
+pub struct Base64DecodeOpts {
+    #[arg(short, long, value_parser = validate_path, default_value = "-")]
+    pub input: String,
+    #[arg(long, value_parser = parse_base64_format, default_value = "standard")]
+    pub format: Base64Format,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub enum Base64Format {
+    Standard,
+    UrlSafe,
+}
+
+fn parse_base64_format(format: &str) -> Result<Base64Format, anyhow::Error> {
+    format.parse()
+}
+
+impl FromStr for Base64Format {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "standard" => Ok(Base64Format::Standard),
+            "urlsafe" => Ok(Base64Format::UrlSafe),
+            _ => Err(anyhow::anyhow!("Invalid format")),
+        }
+    }
+}
+
+impl From<Base64Format> for &'static str {
+    fn from(value: Base64Format) -> Self {
+        match value {
+            Base64Format::Standard => "standard",
+            Base64Format::UrlSafe => "urlsafe",
+        }
+    }
+}
+
+impl fmt::Display for Base64Format {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", Into::<&str>::into(*self))
+    }
+}