+++ src/main.rs
@@ -1,11 +1,14 @@
-mod opts;
-
-use crate::opts::{Opts, SubCommand};
+use base64::engine::general_purpose::URL_SAFE;
+use base64::prelude::BASE64_STANDARD;
+use base64::Engine;
 use clap::Parser;
 use csv::Reader;
 use rand::Rng;
+use rust_cli::base64::{Base64Format, Base64SubCommand};
+use rust_cli::opts::{Opts, SubCommand};
 use serde_json::Value;
 use std::fs;
+use std::io::stdin;
 
 /// rust-li csv -i input.csv -o output.json -d ","
 fn main() -> anyhow::Result<()> {
@@ -37,6 +40,46 @@ fn main() -> anyhow::Result<()> {
             println!("Password: {}", pass);
             Ok(())
         }
+        SubCommand::Base64(base64_sub_command) => match base64_sub_command {
+            Base64SubCommand::Encode(opts) => {
+                println!("Encoding: {:?}", opts);
+                let mut std_in_str = String::new();
+                stdin().read_line(&mut std_in_str)?;
+                // 优化去掉/r/n
+                let input_str = std_in_str.trim_end();
+                //打印输入的字符串
+                print!("Input: {:?} ", input_str);
+                //编码
+                let mut buf = String::new();
+                match opts.format {
+                    Base64Format::Standard => {
+                        BASE64_STANDARD.encode_string(input_str, &mut buf);
+                    }
+                    Base64Format::UrlSafe => {
+                        URL_SAFE.encode_string(input_str, &mut buf);
+                    }
+                }
+                println!("Encoded: {}", buf);
+                Ok(())
+            }
+            Base64SubCommand::Decode(opts) => {
+                println!("Decoding: {:?}", opts);
+                let mut std_in_str = String::new();
+                stdin().read_line(&mut std_in_str)?;
+                // 优化去掉/r/n
+                let input_str = std_in_str.trim_end();
+                //打印输入的字符串
+                print!("Input: {:?} ", input_str);
+                //编码
+                let decoded = match opts.format {
+                    Base64Format::Standard => BASE64_STANDARD.decode(input_str)?,
+                    Base64Format::UrlSafe => URL_SAFE.decode(input_str)?,
+                };
+                let decoded_str = String::from_utf8(decoded)?;
+                println!("Decoded: {}", decoded_str);
+                Ok(())
+            }
+        },
     }
 }